pyproject.toml
[tool.poetry]
name = "photos2book"
version = "0.1.0"
description = ""
authors = ["vitek <vitekzach@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pillow = "^9.4.0"
tqdm = "^4.64.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
